# Add precompiled benchmarks for Mond to show only runtime performance