Hi there, I made Mond - thanks for setting this up! I went in to check why Mond is so slow and saw that the script parsing and compilation time is included in the benchmarks. This isn't ideal because most real world use cases of scripting languages would leverage cached compilations of scripts when possible.

This PR adds a new set of Mond benchmarks that compile the source script once and then reuse it across all runs. See comparison below:

```
| Method                      | LoopCount | Mean       | Error     | StdDev    | Rank | Gen0   | Gen1   | Allocated |
|---------------------------- |---------- |-----------:|----------:|----------:|-----:|-------:|-------:|----------:|
| MondAlloc                   | 100       | 104.054 us | 1.9105 us | 1.7871 us |    2 | 7.0801 | 1.5869 | 352.52 KB |
| MondPrecompiledAlloc        | 100       |  30.490 us | 0.5598 us | 0.5236 us |    1 | 1.8311 | 0.2441 |  91.39 KB |
|                             |           |            |           |           |      |        |        |           |
| MondCSharpToLang            | 100       |  70.209 us | 0.7220 us | 0.6753 us |    2 | 5.2490 | 0.6104 | 260.46 KB |
| MondPrecompiledCSharpToLang | 100       |   6.617 us | 0.1320 us | 0.1296 us |    1 | 0.4654 |      - |  22.96 KB |
|                             |           |            |           |           |      |        |        |           |
| MondLangToCSharp            | 100       |  78.704 us | 1.4717 us | 1.3766 us |    2 | 5.4932 | 0.4883 | 271.41 KB |
| MondPrecompiledLangToCSharp | 100       |   8.487 us | 0.1691 us | 0.1809 us |    1 | 0.3052 |      - |  15.11 KB |
```

Mond's compiler adds a lot of overhead! I never bothered to optimize it and it shows. Doesn't help that it compiles from an `IEnumerable<char>` instead of a flat buffer. Can do cool stuff with that though - check out MondKeyboard.

I am not familiar with the other benchmarked libraries but they should all have similar functionality to only compile once. I think it's worth adding similar benchmarks for everything else.