@Veslo5 2025-01-30
Hi, Rohansi. First off, Mond is awesome library with good API and great syntax!

I've implemented Mond by looking on repo wiki and maybe I overlooked part about script precompiling. Anyway if the library offers it, I primarily tried to reuse the state(VM) and precompile scripts. I'll try to optimize usage of libraries along the way.

Thanks for your work on Mond and this PR! 
Will merge after work.