@Rohansi 2025-01-30
Ah, I see Jint is already compiling the scripts in `Setup`! But it is allocating a new `Engine` in the benchmarks because previous runs would redeclare variables. I believe it should be possible to work around that issue by wrapping all of the JavaScript scripts in `return function() { ... };` and cache the function `JsValue` instead of the precompiled script. Then you can use `JsVM.Call` in the benchmarks without needing a new `Engine` instance each run.